+++ src/scripts/python_project_packages_install.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+# ******************************************************************************
+# A bash script to install Python modules in the project environment required
+# by the project.
+# ******************************************************************************
+pip install --upgrade \
+  pip \
+  setuptools