+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 ### HEAD
 
-### 0.0.2 (April 21, 2020)
+### 0.0.2 (April 25, 2020)
 
   * Updated the copyright date,
   * Made a few small improvements,
   * Added a README explaining how to install and setup Jupyter for the project,
   * Updated createpkg.sh script,
+  * Added the script configure.sh,
   * ...,